+++ NEWS.md
@@ -1,4 +1,4 @@
-# paletteer (development version)
+# paletteer 1.1.0
 
 * added `scale_color_paletteer_binned()` to support binned scales.
 * 20 palettes have been added from the futurevisions package. (82)